- fix gyroscope

+++ app/src/main/java/com/jamnguyen/spirit/XConfig.java
@@ -7,7 +7,7 @@ public class XConfig
     public static final boolean     isTEAM_STORMX               = false;
     public static final boolean     USE_ROTATION_VECTOR         = false;
     public static final boolean     USE_GYROSCOPE               = true;
-    public static final double     GYROSCOPE_MOMENT_DEFAULT               = 86;//Phone của Tiến
+    public static final int     GYROSCOPE_MOMENT_DEFAULT               = 86;//Phone của Tiến
     public static final int     DEGREE_DELTA               = 30;
     public static final int     DEGREE_DELTA_SMALL               = 15;
     public static final long     TIME_REMAIN_FOR_BACK_GOAL = 60*1000;
